Move `PartialGlobalVertex` to `partial` module
