+++ crates/fj-kernel/src/builder/mod.rs
@@ -7,7 +7,6 @@ mod face;
 mod shell;
 mod sketch;
 mod solid;
-mod vertex;
 
 pub use self::{
     curve::{CurveBuilder, GlobalCurveBuilder},
@@ -17,5 +16,4 @@ pub use self::{
     shell::ShellBuilder,
     sketch::SketchBuilder,
     solid::SolidBuilder,
-    vertex::PartialGlobalVertex,
 };
